services/glpn_depth/requirements.txt
transformers
torch
torchvision
Pillow
opencv-python